+++ packages/xchain-wallet/src/index.ts
@@ -7,3 +7,5 @@ import { Wallet } from './wallet'
  * Export Wallet module
  */
 export { Wallet }
+
+export { ChainBalances, EvmTxParams, UtxoTxParams } from './types'